Polishing and code simplification
- LogFileFilterBuildWrapper changed into LogFileFilterConfig; now it
extends Describable instead of SimpleBuildWrapper. This simplifies the
code and removes the useless checkbox "Log File Filter" in the project
configuration view.

With this name change the folder in which the .jelly files lie is also
changed accordingly. Nevertheless, nothing has been modified in the
jelly files.

- Added try/catch to LogFileFilterOutputStream constructor so that
findBugs stops complaining.